% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgData.R
\docType{data}
\name{AgData}
\alias{AgData}
\title{Simulated Agriculure Study}
\format{
A data.frame with 96 observations
\describe{
    \item{plot}{Which plot the observation comes from}
    \item{subplot}{Which subplot the observation comes from}
    \item{rep}{The replicate number within subplot}
    \item{Irrigation}{The irrigation level of the plot}
    \item{Fertilizer}{The fertilizer level of the subplot}}
}
\usage{
AgData
}
\description{
This dataset describes a ficticous split-plot agriculture study in which we have
eight plots where we can only apply one of two irrigation level (low/high).  
Within each plot, we have four sub-plots where we can apply a fertilizer 
treatment and we have two fertilizer treatments (low/high).  Within each 
sub-plot we have five replicate samples.
}
\keyword{datasets}
