% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.anthro.R
\docType{data}
\name{anthro2}
\alias{anthro2}
\title{Simulated dataset ANTHRO 2}
\description{
Simulated dataset ANTHRO 2, in a data.frame with 796 observations of 5 harmonized
variables. The ANTHRO dataset contains synthetic data based. This dataset does not
contain any NA values.
}
\details{
\tabular{llll}{
   \strong{Variable} \tab \strong{Description} \tab \strong{Type} \tab \strong{Note} \cr
   age \tab Age of child \tab numeric \tab months \cr
   sex \tab Gender of child \tab integer \tab  \cr
   weight \tab Weight of child \tab numeric \tab kg \cr
   height \tab Height of child \tab numeric \tab cm \cr
   muac \tab Mid-upper arm circumference \tab numeric \tab cm \cr
}
}
\keyword{data}
