% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arkansas.R
\docType{data}
\name{Arkansas}
\alias{Arkansas}
\title{Time Series of Macroinvertabrates Abundance in the Arkansas River.}
\format{A data frame with 90 observations on the following 2 variables.
\describe{
  \item{year}{The year of observation}
  \item{sqrt.abundance}{The Square root of observed abundance.}
}}
\usage{
Arkansas
}
\description{
A time series of 16 years (5 replicates per year) of mayfly 
(Ephemeroptera:Heptageniidae) abundance in the fall at the monitoring station AR1
on the Arkansas River in Colorado, USA. These data were taking from the R 
package SiZer.
}
\keyword{datasets}
