% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Babies.R
\docType{data}
\name{Babies}
\alias{Babies}
\title{Simulated data set of baby growth}
\format{
a data frame with 78 observations 
\describe{
  \item{Age}{The age in months the observation was made at.}
  \item{Baby}{A number identifying which baby the observation was on.}
  \item{Weight}{The weight (in pounds) of the baby.}}
}
\usage{
Babies
}
\description{
This dataset is an example of a repeated measures study
where there are weight measurements every two months on
six babies from 12 to 36 month.
}
\examples{
library(ggplot2)
ggplot(Babies, aes(x=Age, y=Weight, color=Baby)) +
  geom_point() +
  geom_line(aes(group=Baby))
}
\keyword{datasets}
