% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PainStudy.R
\docType{data}
\name{BackPain}
\alias{BackPain}
\title{Simulated data on the treatment of back pain}
\format{
A data frame with 72 rows and 3 variables:
\describe{
  \item{Electro}{The medical treatment: either None, Low, or High, where None
  is corresponds to no therapy.} 
  \item{Lifestyle}{Change in Lifestyle: either Passive, PT, or Str. These 
  correspond to no change, Physical Therapy, or Strength training.}
  \item{delta}{The reported change in chronic pain.}
}
}
\usage{
BackPain
}
\description{
Chronic back pain can be a debilative problem that causes people to reduce
their motion which can lead to further problems. As such addressing chronic
back pain as a way of keeping the rest of the body healthy is of considerable 
concern.
}
\details{
This data set simulates an experiment in which we consider
three levels of electrotherapy intervention (No Electrotherapy, electrotherapy using 
low frequency simulation, electrotherapy using high frequency, where the low and
high are describing the electrical simulation in Hz). A secondary type of 
treatment regarding lifestyle changes (None, Physical Therapy, Strength Training)
was also included. At the end of 6 months of therapy, patients
rated their observed increase or decrease in pain levels.
}
\keyword{datasets}
