% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeeDrink.R
\docType{data}
\name{BeeDrink}
\alias{BeeDrink}
\title{Drinking energetics of honeybees.}
\format{A data frame with 24 rows and 3 variables:
\describe{
  \item{temp}{Temperature (in Celcius).}
  \item{sucrose}{Percent sucrose of the liquid where higher sucrose levels correspond to higher viscosity.}
  \item{energy}{The energy expended during drinking as joules/second.}
}}
\usage{
BeeDrink
}
\description{
An entomologist conducted an experiment to understand the how 
much energy a honeybee expends while drinking liquids at different
temperatures and different viscosity of liquid. To control viscosity,
the researcher varied the amount of sucrose in the liquid (20,40 or 60%).
}
\details{
The data was provided from Dr. S. Buckman, USDA Bee Research Lab, Tucson, AZ.
}
\keyword{datasets}
