% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bird.Detections.R
\docType{data}
\name{Bird.Detections}
\alias{Bird.Detections}
\title{Bird Dectections - Simulated Occupancy data}
\format{A matrix with 40 rows (sites) and 5 columns (observation
periods).}
\usage{
Bird.Detections
}
\description{
A simulated dataset of a occupancy study of birds where we
observe 40 different sites (rows) at 5 different observation
time periods. The study undertaken during 10 days during the
after the start of the breeding season (so there should be little 
to no immigration/emmigration). Each site was visited every
other day to obtain independece between successive sampling 
events.
}
\keyword{datasets}

