% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainBody.R
\docType{data}
\name{BrainBody}
\alias{BrainBody}
\title{Brain vs Body Size}
\format{A data frame with 76 rows and 9 variables:
\describe{
  \item{Species}{Species Name}
  \item{Brain}{Average brain weight (g)}
  \item{Body}{Average body weight (kg)}
  \item{Gestation}{Average gestation length (days)}
  \item{Litter}{Average litter size}
  \item{LitterSize}{Discritized litter size }
}}
\usage{
BrainBody
}
\description{
A set of 96 mammel species.
}
\keyword{datasets}
