% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cats.R
\docType{data}
\name{Cats}
\alias{Cats}
\title{Cat Behaviors}
\format{A data frame with 60 rows and 3 variables:
\describe{
  \item{Temperature}{Outside temperature (in Celcius).}
  \item{Time}{The amount of time spent outside (in minutes).}
  \item{Cat}{A categorical variable denoting which cat was outside.}
  }}
\usage{
Cats
}
\description{
A simulated dataset describing how my time my cats (Beau, Tess, and Kaylee) spend outside
relative to the outside temperature.
}
\keyword{datasets}
