% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChildGrowth.R
\docType{data}
\name{ChildGrowth}
\alias{ChildGrowth}
\title{Child Growth Dataset}
\format{
A data frame with growth curves from 10 male and 10 females. 
\describe{
  \item{childID}{A unique identify for each child.}
  \item{sex}{The sex for each child.}
  \item{age}{The child's age at the observation.}
  \item{height}{The child's height in centimeters.}
}
}
\usage{
ChildGrowth
}
\description{
This is a simulated data set of child growth curves from 4 to 12 years of
age. This utilizes the a United States CDC data of growth curves to sample from
using the `childsds` package.
}
\examples{
library(ggplot2)
ggplot(ChildGrowth, aes(x=age, y=height)) +
  geom_point() +
  geom_line(aes(group=childID)) +
  facet_grid(.~sex)
}
\keyword{datasets}
