% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommunitySurvey.R
\docType{data}
\name{CommunitySurvey}
\alias{CommunitySurvey}
\title{A simulated survey of comparing the species make-up of two different
ecological communities.}
\format{
A data frame with 34 rows and 6 variables:
\describe{
  \item{Community}{A factor denoting what community type the observation came from}
  \item{A}{The number species A observed.}
  \item{B}{The number species B observed.}
  \item{C}{The number species C observed.}
  \item{D}{The number species D observed.}
  \item{E}{The number species E observed.}
}
}
\usage{
CommunitySurvey
}
\description{
This data is a simulation of a survey designed to measure the difference
in ecological diversity of two different communities. Each row represents a 
site and we have observed the abundances of 5 different species (labeled A, B, ..., E).
The two community types are labeled I and II.
}
\examples{
library(tidyverse)
library(vegan)
data('CommunitySurvey', package='dsData')
DistMatrix <- CommunitySurvey \%>\% 
  select(-Community) \%>\%
  vegdist( )
model <- anosim(DistMatrix, CommunitySurvey$Community)
summary(model)
model$statistic
}
\keyword{datasets}
