% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Polishing.R
\docType{data}
\name{Dishes}
\alias{Dishes}
\title{Manufacturing data on how long it takes to polish a dish}
\format{A data frame with 58 rows and 4 variables:
\describe{
  \item{Diameter}{The diameter of the dish.}
  \item{Time}{The amount of time necessary to polish the dish.}
  \item{Price}{How much the dish will sell for.}
  \item{Type}{The type of dish: Casserole, Bowl, Dish, Tray, or Plate}
}}
\usage{
Dishes
}
\description{
This data describes the final step in processing dishes
before the are packaged and sold. The polishing time could
depend on the size of the dish, the price that it will be sold
at, and the type of dish.
}
\keyword{datasets}

