% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Eden.R
\docType{data}
\name{Eden}
\alias{Eden}
\title{An elevation map of Maunga Whau (Mt Eden).}
\format{A data frame with 4800 rows and 5 variables:
\describe{
  \item{x}{the x coordinate}
  \item{y}{the y coordinate}
  \item{elevation}{the elevation (in meters)}
  \item{PSU}{the primary sampling unit associated with an observation}
  \item{SSU}{the secondary sampling unit associated with an observation. These are 1-25 and nested within PSU}
}}
\usage{
Eden
}
\description{
Maunga Whau (Mt Eden) one of about 50 volcanos
in the Auckland volcanic field. This dataset takes the volcano dataset and
adds a primary sampling unit (PSU) and secondary sampling unit (SSU) onto
the data so that it is suitable for cluster sampling examples.
}
\examples{
# Plot the elevation
ggplot(Eden, aes(x=x, y=y, fill=elevation)) +
     geom_tile()

# Plot the PSUs
PSU_labels <- Eden \%>\% group_by(PSU) \%>\% summarize(x=mean(x), y=mean(y), label=mean(PSU))
ggplot(Eden, aes(x=x, y=y, fill=(PSU-1) \%/\% 16)) +
    geom_tile() +
    geom_text( data=PSU_labels, aes(label=label))

# PSUs modulus 4 and 16 are geographically contiguous
Eden.PSU <- Eden \%>\% group_by(PSU) \%>\%
  summarize(x=mean(x), y=mean(y))
ggplot(Eden, aes(x=x, y=y, fill=(PSU-1)\%/\%16)) +
  geom_tile() +
  geom_path(data=Eden.PSU)
}
\keyword{datasets}

