% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlagstaffTemp.R
\docType{data}
\name{FlagstaffTemp}
\alias{FlagstaffTemp}
\title{Daily maximum termperature in Flagstaff, AZ}
\format{
A data frame with 5 variables:
\describe{
  \item{Date}{The date of the observation.}
  \item{Year}{Year of the observation.}
  \item{Month}{Month of the observation.}
  \item{Day}{Day of month of the observation.}
  \item{MaxTemp}{Daily maximum temperature (in degrees Farenheit).}
}
}
\usage{
FlagstaffTemp
}
\description{
Information about the daily maximum temperture in Flagstaff, AZ from May 1985 to December 2015.
}
\keyword{datasets}
