% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golftees.R
\docType{data}
\name{golftees}
\alias{golftees}
\title{Golftee line transect data}
\format{A data frame with ...
  \describe{
    \item{column1}{description}
    \item{column2}{description}
  }}
\source{
Data provided courtesy of ...
}
\description{
A brief 1 paragraph explanation of the data
}
\details{
More details
}
\references{
A paper/report
}
\keyword{datasets}
