% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalData.R
\docType{data}
\name{HierarchicalData}
\alias{HierarchicalData}
\title{A simulated hierarchical analysis}
\format{A data frame with 54 rows and 4 variables:
\describe{
  \item{Trt}{A categorical variable (A,B,C) denoting what treatment was applied.}
  \item{Ring}{A categorical varible (1,...,9) denoting which ring the observation was from.}
  \item{Rep}{A number from (1,2,...,6) denoting which replicate within a ring.}
  \item{y}{The response variable.}
}}
\usage{
HierarchicalData
}
\description{
This data is simulated an experiment that has some 
hierachical nesting and the treatment is not applied at
the observation level, but rather at some higher blocking level. 
This is a plant response experiment where there are 9 rings, three of which
are randomly assigned to treatment A, three assigned to 
treatment B and the final 3 to treatment C. Within each
ring, we observe six individual plants. So we have 6*9=54
observations, but the plants are nested within the ring.
}
\examples{
library(ggplot2)
ggplot(HierarchicalData, aes(y=y, x=Ring, color=Trt)) +
  geom_point()
}
\references{
This dataset was inspired by: Evans et al (2014) Greater ecosystem carbon in the Mojave Desert 
after ten years exposure to elevated CO2. Nature Climate Change 4(5): 394-397.
}
\keyword{datasets}

