% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Humpback.R
\docType{data}
\name{Humpback}
\alias{Humpback}
\title{Simulated distance sampling of whales.}
\format{An object of class \code{data.frame} with 36 rows and 2 columns.}
\usage{
Humpback
}
\description{
A simulated dataset of a distance transect study
of Humpback whales during the summer feeding season.
The design had 25 different transects (all of length 
1000 meters), that were systematically spread across a 
study area that covered 100 square kilometers. 
The recorded distances are distance from the individual
to the transect in meters.
}
\details{
#' @format A data frame with 36 rows and 2 variables:
\describe{
  \item{Sample.Label}{Which transect was the observation made on.}
  \item{distance}{Distance to the transect, in meters.}
}
}
\keyword{datasets}

