% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IceCream.R
\docType{data}
\name{IceCream}
\alias{IceCream}
\title{Ice Cream sales versus Average Temperture}
\format{
A dataframe with 12 rows and 3 columns
\describe{
  \item{Month}{The month.}
  \item{Ave_Temp}{Average daily maximum (in degrees Fahrenheit).}
  \item{Sales}{Ice creams sales (in 1000s of gallons).}
  }
}
\source{
https://www.businessinsider.com/pop-quiz-guess-the-1-month-for-ice-cream-production-in-america-2012-3
}
\usage{
IceCream
}
\description{
The ice cream sales are taken from the Department of Agriculture
via a article from "Business Insider". The data is from the year 2011.
}
\keyword{datasets}
