% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/LichenGrowth.R
\docType{data}
\name{LichenGrowth}
\alias{LichenGrowth}
\title{Simulated Lichen Growth}
\format{A data frame with 36 rows and 2 variables:
\describe{
  \item{Species}{Lichen Species}
  \item{Water}{A normalized measure of water availability (bigger is more water)}
  \item{Growth}{A measure of how much growth occured (bigger is more growth)}
}}
\usage{
LichenGrowth
}
\description{
At 20 different site, 100 seedlings were planted. Two watering treatments
}
\details{
#'
}
\keyword{datasets}

