% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lobsters.R
\docType{data}
\name{Lobsters}
\alias{Lobsters}
\title{A simulated dataset of lobster captures}
\format{A data frame with 240 rows and 4 variables:
\describe{
  \item{Boat.ID}{A variable denoting which boat.}
  \item{Catch.Weight}{The weight of all lobsters caught (by boat, in thousands of pounds).}
  \item{Lobster.ID}{A variable indicating which lobster.}
  \item{Lobster.Length}{Length of lobsters carapace.}
}}
\usage{
Lobsters
}
\description{
This data is a simulation of a multistage sample. The given
scenario is that 31 boats enter a harbor on a given day and
12 different boats or selected for inspection. Aboard each boat,
20 lobsters are sampled and the length of the carapace is measured.
Note that any lobster with carapace length less than 3 inches
is legally required to be released unharmed at its capture site.
}
\examples{
data(Lobsters)
# Plot each boats Catch.Weight 
ggplot(Lobsters, aes(x=Boat.ID, y=Catch.Weight)) + 
  geom_point()
     
# Plot the lobster lengths, by boat
ggplot(Lobsters, aes(x=Boat.ID, y=Lobster.Length)) +
  geom_point()    
ggplot(Lobsters, aes(x=factor(Boat.ID), y=Lobster.Length)) +
  geom_boxplot()
     
# Plot Lobster Length versus the total Catch Weight
ggplot(Lobsters, aes(x=Catch.Weight, y=Lobster.Length)) +
  geom_point()
}
\keyword{datasets}

