% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mayflies.R
\docType{data}
\name{Mayflies}
\alias{Mayflies}
\title{Mayfly Abundance}
\format{
A data frame with 30 observations, with two columns 
\describe{
   \item{CCU}{Heavy metal pollution index (higher is more pollution).}
   \item{Occupancy}{Whether Mayflies were present at the site.}}
}
\usage{
Mayflies
}
\description{
Because mayflies are sensitive to metal pollution, we might be interested in 
looking at the presence/absence of mayflies in a stream relative to a pollution 
gradient. This simulated dataset gives a pollution gradient is measured in 
Cumulative Criterion Units (CCU), where  CCU is defined as the ratio of the 
measured metal concentration to the hardness adjusted chronic criterion 
concentration, and then summed across each metal. Large values of CCU
imply more metal pollution.
}
\references{
This dataset was inspired by: Sonderegger, D.L., Wang, H., 
Clements, W.H., and Noon, B.R. (2009) Using SiZer to detect thresholds in 
ecological data. Frontiers in Ecology and the Environment 7:190-195.
}
\keyword{datasets}
