% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mosquitos.R
\docType{data}
\name{Mosquitos}
\alias{Mosquitos}
\title{Simulated data on Mosquito Abundance}
\format{A data frame with ??? rows and ??? variables:
\describe{
  \item{Abundance}{A measure of mosquito abundance.} 
  \item{MaxSummerTemp}{Maximum Summer Temperature (Farenheit).}
  \item{AnnualPrecip}{Annual Precipitation (in inches).}
  \item{MedianIncome}{Household Median Income.}
  \item{Population}{Population size.}
}}
\usage{
Mosquitos
}
\description{
Mosquitos are a very common disease vector and understanding the climactic
drivers of large mosquito populations will be helpful to understand disease 
outbreaks and how disease might change in the face of climate change.
}
\details{
This data set simulates an observational study where we observe a number 
of communities and record an abundance measurement along with pertinent 
covariates.
}
\keyword{datasets}

