% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oysters.R
\docType{data}
\name{Oysters}
\alias{Oysters}
\title{A simulated dataset of oyster catch}
\format{
A data frame with 21 rows and 2 variables:
\describe{
  \item{Density}{Number of boat passes through the oyster bed.}
  \item{Harvest}{The number of bushels of oysters harvested.}
}
}
\usage{
Oysters
}
\description{
This data describes an oyster harvest. Each observation corresponds to
an hectare of oyster beds and the density describes the intensity in which
the bed was harvested (number of passes by a havesting boat) and the Harvest is 
the number of bushels obtained.
}
\keyword{datasets}
