% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plant.Pop.R
\docType{data}
\name{Plant.Pop}
\alias{Plant.Pop}
\title{A simulated dataset of a population of plants.}
\format{
A data frame with 10,000 rows and 6 variables:
\describe{
  \item{x}{The x coordinate of a 1x1 grid cell.}
  \item{y}{The y coordinate of a 1x1 grid cell.}
  \item{Quadrant}{Which 25x25 quadrant the grid cell is in [a-p].}
  \item{Plot}{Which 10x10 plot the grid cell is in [1-100].}
  \item{Site}{A unique identifier of the 1x1 grid cell [1 - 10,000].}
  \item{count}{The number of plants within a site.}
}
}
\usage{
Plant.Pop
}
\description{
The plants are in a 100 x 100 grid and display
a moderate amount of clustering. The grid is 
divided into 16 25x25 Quadrats, or 100 10x10 Plots.
The data is give as the number of individual plants
in a 1x1 grid cell.
}
\examples{
library(ggplot2)
library(dplyr)
data(Plant.Pop)

NonZero <- Plant.Pop \%>\% filter( count >= 1 )
ggplot( NonZero, aes(x=x, y=y, color=count)) + 
  geom_point(alpha=.3)

# Aggregate over quadrants
Plant.Quads <- Plant.Pop \%>\%
  group_by(Quadrant) \%>\%
  summarize(x=mean(x),
            y=mean(y),
            count=sum(count))   
ggplot(Plant.Quads, aes(x=x, y=y)) +
  geom_point(data=Plant.Pop, alpha=.3, aes(color=count)) +
  geom_text(aes(label=Quadrant))  
  
# Aggregate over plots
Plant.Plots <- Plant.Pop \%>\%
  group_by(Plot) \%>\%
  summarize(x=mean(x),
            y=mean(y),
            count=sum(count))   
ggplot(Plant.Plots, aes(x=x, y=y)) +
  geom_point(data=Plant.Pop, alpha=.3, aes(color=count)) +
  geom_text(aes(label=Plot))  
}
\keyword{datasets}
