% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/PlantNutrients.R
\docType{data}
\name{PlantNutrients}
\alias{PlantNutrients}
\title{A simulated study the effects of soil nutrients on harvest levels}
\format{A data frame with 50 rows and 4 variables:
\describe{
  \item{Harvest}{The number of bushels harvested for the plot}
  \item{N}{Nitrogen concentration (in ppm)}
  \item{P}{Phosporus concentration (in ppm)}
  \item{K}{Potassium concentration (in ppm)}
}}
\usage{
PlantNutrients
}
\description{
The study is an observational study looking at the Nitrogen,
Phosporus, and Potassium levels in 50 plots and along with the
harvest amounts for those plots.
}
\keyword{datasets}

