% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pollution.R
\docType{data}
\name{Pollution}
\alias{Pollution}
\title{A simulated dataset describing the abundance of certain insects
that are thought to be sensitive to heavy metal pollution.}
\format{A data frame with :
\describe{
  \item{Pollution}{Categorical variable denoting one of four levels of pollution.}
  \item{Species}{What species is being observed.}
  \item{Log.Abundance}{The log count of the number of individuals observed.}
}}
\usage{
Pollution
}
\description{
A simulated dataset describing the abundance of certain insects
that are thought to be sensitive to heavy metal pollution.
}
\keyword{datasets}
