% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ponderosa.R
\docType{data}
\name{Ponderosa}
\alias{Ponderosa}
\title{A simulated forestry survey}
\format{
A data frame with 227 rows and 2 variables:
\describe{
  \item{Plot.ID}{A variable denoting which Plot the observation came from}
  \item{Age}{Age of the observed tree.}
}
}
\usage{
Ponderosa
}
\description{
This data is a simulation of a survey designed to estimate the average
age of mature Ponderosa Pine trees. The idea is that 50 randomly selected
0.25 hectare plots are selected and all the mature trees in the plot are
sampled for age (via tree cores).
}
\examples{
data(Ponderosa)
library(ggplot2)

ggplot(Ponderosa, aes(x=Age)) +
  geom_histogram(aes(y=..density..)) +
  geom_density(color='red', size=2)
  
# calculate some summary statistics   
library(dplyr)
Ponderosa \%>\% group_by(Plot.ID) \%>\%
  summarize(mean = mean(Age),
             sd   = sd(Age))
}
\keyword{datasets}
