% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sikadeer.R
\docType{data}
\name{sikadeer}
\alias{sikadeer}
\alias{sikadeer_units}
\title{Sika deer pellet data from southern Scotland, described in}
\format{A data frame with ...
  \describe{
    \item{column1}{description}
    \item{column2}{description}
  }}
\source{
Data provided courtesy of ...
}
\description{
A brief 1 paragraph explanation of the data
}
\details{
More details
}
\references{
Marques, F. F. C., S. T. Buckland, D. Goffin, C. E. Dixon, D. L. Borchers, B. A. Mayle, and A. J. Peace. 2001. Estimating deer abundance from line transect surveys of dung: sika deer in southern Scotland. Journal of Applied Ecology 38 (2): 349–363. https://doi.org/10.1046/j.1365-2664.2001.00584.x
}
\keyword{datasets}
