% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toads.R
\docType{data}
\name{Toads}
\alias{Toads}
\title{A simulated dataset of the presence/absense of boreal toads.}
\format{
A data frame with 200 observations of 4 variables.
\describe{
  \item{Occupancy}{A zero or 1 where a 1 denotes that the site was occupied by toads.}
  \item{Elevation}{The site elevation (in feet).}
  \item{Shore.Veg}{A factor with three levels (Low, Medium, High)}
  \item{Depth}{The maximum depth of the water source (in feet).}}
}
\usage{
Toads
}
\description{
This simulated dataset imagines a study that looks at the 
200 high elevation ponds and lakes and observes if they 
water source was used for breeding by boreal toads. For
each water source, we have information about the elevation,
level of shore vegetation and the maximum depth of the 
pool.
}
\keyword{datasets}
