% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univesity.R
\docType{data}
\name{University}
\alias{University}
\title{Simulated data set of baby growth}
\format{a data frame with 182 observations 
\describe{
  \item{Socio}{The socio-economic background the individual was raised in (Low, Medium, High).}
  \item{Selectivity}{The selectivity of the college the person attended. (Medium, High)}
  \item{Income}{The individuals yearly income, 10 years out of college.}}}
\usage{
University
}
\description{
This dataset is an example of a repeated measures study
where there are weight measurements every two months on
six babies from 12 to 36 month.
}
\keyword{datasets}
