% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/UUnion.R
\docType{data}
\name{UUnion}
\alias{UUnion}
\title{A simulated survey given to NAU students.}
\format{A data frame with 1100 rows and 2 variables:
\describe{
  \item{Year}{A factor denoting what year in school the respondant is.}
  \item{NumPerWeek}{The number of times per week the repondant visits the UU.}
}}
\usage{
UUnion
}
\description{
This data is a simulation of a survey designed to measure how
frequently students visit the University Union Building on campus.
The data collected for each student includes their year in school
(freshman through graduate student) as well as their average number
of visits per week. Information about NAUs student population
demographics can be found at: http://www4.nau.edu/pair/quickfact.asp
}
\examples{
data(UUnion)
ggplot(UUnion, aes(x=Year, y=NumPerWeek)) +
  geom_boxplot()

# calculate some summary statistics
library(dplyr)
UUnion \%>\% group_by(Year) \%>\%
  summarize(mean = mean(NumPerWeek),
            sd   = sd(NumPerWeek))
}
\keyword{datasets}

