% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vehicles.R
\docType{data}
\name{vehicles}
\alias{vehicles}
\title{Automobile fuel economy data from the EPA for the 2015 model year.}
\format{
A data frame with :
\describe{
  \item{make}{Manufacturer}
  \item{model}{Model name}
  \item{class}{Vehicle size class}
  \item{drive}{Drive Train Type}
  \item{cyl}{Number of cylinders}
  \item{displ}{Engine Displacement}
  \item{fuel}{Required Fuel Type}
  \item{year}{Model year}
  \item{Efficient}{A binary response indicating if the the model has a higher city MPG than the 2015 median value. A 1 is higher than the median, a 0 is less.}
}
}
\usage{
vehicles
}
\description{
This dataset contains selected varaibles,  and removes vehicles with 
incomplete data (e.g. no drive train data). These data are a subset of the
package fueleconomy, with only the 2015 model year and the vehicle class
has been simplified.  The column Efficiency was calculated by calculating 
median City MPG and evaluating if a particular model was in the upper or lower
half of the City MPG.
}
\keyword{datasets}
