% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wren.R
\docType{data}
\name{wren}
\alias{wren}
\alias{wren1}
\alias{wren2}
\alias{wren3}
\alias{wren4}
\alias{wren1_units}
\alias{wren2_units}
\alias{wren3_units}
\alias{wren4_units}
\title{Steve Buckland's winter wren surveys}
\source{
Steve Buckland
}
\description{
Observations of winter wren (Troglodytes troglodytes L.) collected by Steve Buckland in woodland/parkland at Montrave Estate near Leven, Fife, Scotland.
}
\details{
Four different surveys were carried out:
\describe{
  \item{\code{wren1}}{5-minute point count}
  \item{\code{wren2}}{snapshot method}
  \item{\code{wren3}}{cue count}
  \item{\code{wren4}}{line transect data}
}
}
\references{
Buckland, S. T. 2006. Point-transect surveys for songbirds: robust methodologies. The Auk 123 (2): 345–357.
}
\keyword{datasets}
