% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zombies.R
\docType{data}
\name{Zombies}
\alias{Zombies}
\title{A simulated dataset about zombies solving a maze.}
\format{A data frame with 39 observations of 4 variables.
\describe{
  \item{Zombie_ID}{A unique identifier for each zombie.}
  \item{Sex}{A factor describing the sex, Male or Female.}
  \item{Age}{The number of days since the zombie was first infected.}
  \item{Time}{The amount of time in the maze.}}}
\usage{
Zombies
}
\description{
This simulated dataset imagines a study where captured zombies are presented
with a maze with the researchers goal to measure the zombies remaining intelligence. 
We imagine that foreach zombie, we know the time since infection as well as the 
zombies sex.
}
\details{
The zombie is released and we measure the time until the zombie reaches the 
treat in the center of the maze. We expect that zombies that have been infected 
for longer will be dumber and move more slowly.
}
\keyword{datasets}
