% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capercaillie.R
\docType{data}
\name{capercaillie}
\alias{capercaillie}
\alias{capercaillie_units}
\title{Capercaillie in Monaughty Forest}
\description{
Data from a line transect survey of capercaillie in Monaughty Forest, Moray, Scotland.
}
\keyword{datasets}
