% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducknests.R, R/unimak.R
\docType{data}
\name{ducknests}
\alias{ducknests}
\alias{ducknests_units}
\title{Ducknest line transect survey data}
\format{A data frame with ...
  \describe{
    \item{column1}{description}
    \item{column2}{description}
  }}
\source{
Simulated data, from the distance sampling introductory course, Centre for Research into Ecological & Environmental Modelling, University of St Andrews.

Simulated data, from the distance sampling introductory course, Centre for Research into Ecological & Environmental Modelling, University of St Andrews.
}
\description{
Simulated line transect survey of duck nests, designed to reproduce the data of Figure 2 in Anderson and Pospahala (1970).

Simulated line transect survey of duck nests, designed to reproduce the data of Figure 2 in Anderson and Pospahala (1970).
}
\references{
Anderson, D. R., and R. S. Pospahala. 1970. Correction of bias in belt transect studies of immotile objects. The Journal of Wildlife Management 34 (1): 141–146. https://doi.org/10.2307/3799501

Anderson, D. R., and R. S. Pospahala. 1970. Correction of bias in belt transect studies of immotile objects. The Journal of Wildlife Management 34 (1): 141–146. https://doi.org/10.2307/3799501
}
\keyword{datasets}
