% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrozenShoulder.R
\docType{data}
\name{Frozen}
\alias{Frozen}
\title{Simulated data on treatment of frozen shoulder.}
\format{An object of class \code{data.frame} with 45 rows and 3 columns.}
\usage{
Frozen
}
\description{
Adhesive capsulitis (also known as Frozen shoulder) is a painful and 
disabling disorder of unclear cause in which the shoulder capsule, 
the connective tissue surrounding the glenohumeral joint of 
the shoulder, becomes inflamed and stiff, greatly restricting motion 
and causing chronic pain. (from Wikipedia)
}
\details{
This data set simulations an experiment in which we consider
three levels of medical intervention (nothing, 1 round of cortisone shots,
or three rounds of cortisone shots) and three levels of manipulation
(nothing, weekly physical therapy, weekly physical therapy and massage).
At the end of 6 months, the patients increase or decrease in range of
motion was recorded.

#' @format A data frame with 36 rows and 2 variables:
\describe{
  \item{Med}{The medical treatment: either nothing, 1 cortisone shot, or 
  a sequence of three shots over the study period}
  \item{Manip}{The manipulative treatment: either nothing, physical 
  therapy, or physical therapy and massage}
  \item{delta}{The increase in range of motion and the conclusion
  of the study (in degrees)}
}
}
\keyword{datasets}

