% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Joinery.R
\docType{data}
\name{Joinery}
\alias{Joinery}
\title{Testing Cabinet Drawer Joinery Methods (Simulated Data!)}
\format{
A data frame with 36 observations of 3 variables.
\describe{
  \item{Wood}{A factor describing the wood type (Plywood, Pine, Maple)}.
  \item{Type}{A factor describing the joint type.}
  \item{Force}{The force required to break the joint (in pounds).}
  }
}
\usage{
Joinery
}
\description{
One defining distinction between cheap furniture and high quality
furniture is the method in which pieces of wood are joined. For
making cabinet drawers, a range of wood products and joinery methods 
used. Often plywood is used, but softwoods like pine or hardwoods are
also used. For joining the sides of the boxes, we could use just 
glue-and-screw the pieces together using a butt-joint, or we could be 
use other more refined techniques like the box-joint or dovetail joint.
}
\details{
Going into the experiment, it is expected that plywood using glue-and-screw
butt-joints will be poor, but it isn't clear how well the other joint types 
will compare when using plywood vs natural woods.
}
\examples{
ggplot2::ggplot(Joinery, aes(x=Wood, y=Force, color=Type)) + geom_point()
}
\keyword{datasets}
