% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LichenGrowth.R
\docType{data}
\name{LichenGrowth}
\alias{LichenGrowth}
\title{Simulated Lichen Growth}
\format{
A data frame with 60 rows and 3 variables:
\describe{
  \item{Species}{Lichen Species}
  \item{Water}{A normalized measure of water availability (bigger is more water)}
  \item{Growth}{A measure of how much growth occured (bigger is more growth)}
}
}
\usage{
LichenGrowth
}
\description{
A study was performed examining the amount of growth
over a season relative to the amount of water available.
}
\details{
Confession: I didn't spend the time
to get the scale for the water or for growth to 
be biologically correct for how they actually get measured.
}
\keyword{datasets}
