% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univesity.R
\docType{data}
\name{University}
\alias{University}
\title{Simulated data set of baby growth}
\format{a data frame with 182 observations 
\describe{
  \item{Socio}{The socio-economic background the individual was raised in (Low, Medium, High).}
  \item{Selectivity}{The selectivity of the college the person attended. (Medium, High)}
  \item{Income}{The individuals yearly income, 10 years out of college.}}}
\usage{
University
}
\description{
This dataset is simulated data that represents a hypothetical study where
we observe 182 individuals with different socio-economic status but
similar high school GPAs and SAT scores. We then observe the selectivity of 
the college they attend and then finally
their salary 10 years after graduating from college. The hypotheses is that
all else being equal, attending a highly selective college will matter more
for children with a low socio-economic status than for children of high
socio-economic parents.
}
\keyword{datasets}
